+++ modules/ui_common.py
@@ -164,10 +164,10 @@ def create_output_panel(tabname, outdir):
 
         with gr.Column():
             with gr.Row(elem_id=f"image_buttons_{tabname}", elem_classes="image-buttons"):
-                open_folder_button = gr.Button('show', visible=not shared.cmd_opts.hide_ui_dir_config)
-                save = gr.Button('save', elem_id=f'save_{tabname}')
-                save_zip = gr.Button('zip', elem_id=f'save_zip_{tabname}')
-                delete = gr.Button('delete', elem_id=f'delete_{tabname}')
+                open_folder_button = gr.Button('Show', visible=not shared.cmd_opts.hide_ui_dir_config)
+                save = gr.Button('Save', elem_id=f'save_{tabname}')
+                save_zip = gr.Button('Zip', elem_id=f'save_zip_{tabname}')
+                delete = gr.Button('Delete', elem_id=f'delete_{tabname}')
                 buttons = parameters_copypaste.create_buttons(["img2img", "inpaint", "extras"])
 
             open_folder_button.click(fn=lambda: open_folder(shared.opts.outdir_samples or outdir), inputs=[], outputs=[])